scheduler service

+++ src/Service/SchedulerService.php
@@ -53,30 +53,35 @@ class SchedulerService
 	}
 
 	/**
-	 * @param int $jobId
+	 * @param string $jobClassName
 	 * @param int $projectId
 	 * @param int $period
 	 * @param bool $active
+	 * @throws \Exception
 	 */
-	public function setProjectSettings(
-		int $jobId,
+	public function addProject(
+		string $jobClassName,
 		int $projectId,
 		int $period,
 		bool $active = true
 	): void
 	{
-		if ($this->em->getRepository(\Sellastica\Scheduler\Entity\SchedulerProject::class)->findOneBy([
-			'jobId' => $jobId,
-			'projectId IS NULL',
-		])) {
+		//find job by class name
+		if (!$job = $this->findJobByClassName($jobClassName)) {
+			throw new \Exception("Job $jobClassName does not exist");
+		}
+
+		//check if job is set for all projects already
+		if ($this->isForAllProjects($job->getId())) {
 			throw new \Exception("Cannot set job settings for project $projectId. This job is set for all projects already.");
 		}
 
+		//find or create project setting
 		if (!$projectSetting = $this->em->getRepository(\Sellastica\Scheduler\Entity\SchedulerProject::class)->findOneBy([
-			'jobId' => $jobId,
+			'jobId' => $job->getId(),
 			'projectId' => $projectId,
 		])) {
-			$projectSetting = \Sellastica\Scheduler\Entity\SchedulerProjectBuilder::create($jobId, $period)
+			$projectSetting = \Sellastica\Scheduler\Entity\SchedulerProjectBuilder::create($job->getId(), $period)
 				->projectId($projectId)
 				->build();
 			$this->em->persist($projectSetting);
@@ -87,16 +92,49 @@ class SchedulerService
 	}
 
 	/**
-	 * @param int $jobId
+	 * @param string $jobClassName
 	 * @param int $projectId
 	 */
-	public function removeProjectSettings(int $jobId, int $projectId): void
+	public function removeProject(string $jobClassName, int $projectId): void
 	{
+		//find job by class name
+		if (!$job = $this->findJobByClassName($jobClassName)) {
+			throw new \Exception("Job $jobClassName does not exist");
+		}
+
+		//check if job is set for all projects
+		if ($this->isForAllProjects($job->getId())) {
+			throw new \Exception("Cannot remove job settings for project $projectId. This job is set for all projects.");
+		}
+
 		if ($projectSetting = $this->em->getRepository(\Sellastica\Scheduler\Entity\SchedulerProject::class)->findOneBy([
-			'jobId' => $jobId,
+			'jobId' => $job->getId(),
 			'projectId' => $projectId,
 		])) {
 			$this->em->remove($projectSetting);
 		}
+	}
+
+	/**
+	 * @param string $className
+	 * @return \Sellastica\Scheduler\Entity\SchedulerJobSetting|null
+	 */
+	private function findJobByClassName(string $className): ?\Sellastica\Scheduler\Entity\SchedulerJobSetting
+	{
+		return $this->em->getRepository(\Sellastica\Scheduler\Entity\SchedulerJobSetting::class)->findOneBy([
+			'className' => $className,
+		]);
+	}
+
+	/**
+	 * @param int $jobId
+	 * @return bool
+	 */
+	private function isForAllProjects(int $jobId): bool
+	{
+		return $this->em->getRepository(\Sellastica\Scheduler\Entity\SchedulerProject::class)->existsBy([
+			'jobId' => $jobId,
+			'projectId IS NULL',
+		]);
 	}
 }
